v2/services/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
celery==5.5.3
redis==6.4.0
openai==1.107.2
requests==2.32.5
oss2==2.19.1
yt-dlp==2026.2.21
email-validator==2.2.0
alembic==1.16.5
prometheus-client==0.23.1
pytest==8.4.2
